keyhive_codec trait
