+++ keyline/src/lib.rs
@@ -1,8 +1,8 @@
 //! Keyline: Keyhive's convergent authority graph.
 //!
-//! A flat namespace of Ed25519 verifying keys ([`Id`]) and a set of signed
-//! certificates over them: [`Delegation`]s that grant an [`Access`] level over
-//! a subject, and [`Revocation`]s that withdraw a delegation by hash. Authority
+//! A flat namespace of Ed25519 verifying keys ([`id::Id`]) and a set of signed
+//! certificates over them: [`delegation::Delegation`]s that grant an [`access::Access`] level over
+//! a subject, and [`revocation::Revocation`]s that withdraw a delegation by hash. Authority
 //! is reachability over that graph, attenuated to the minimum along a route and
 //! combined as the maximum over routes. Evaluation is a pure function of the
 //! set, so every replica holding the same certificates computes the same
@@ -14,14 +14,14 @@
 //!
 //! # Naming
 //!
-//! _Keyline_ is the design, `keyline` is the crate, [`Keyline`] is the trait.
+//! _Keyline_ is the design, `keyline` is the crate, [`keyline::Keyline`] is the trait.
 //!
 //! # What this crate does not do
 //!
-//! It does not verify signatures (`Keyline::insert` takes a [`Verified`]
+//! It does not verify signatures (`Keyline::insert` takes a [`signed::Verified`]
 //! witness), does not know about prekeys, CGKA, documents, or groups, and is
 //! synchronous: concurrency is the wrapper's job. `keyhive_core` holds an
-//! implementation behind a `RwLock` and converts its typed handles to [`Id`]s
+//! implementation behind a `RwLock` and converts its typed handles to [`id::Id`]s
 //! at the boundary.
 //!
 //! # `no_std` support
@@ -49,12 +49,3 @@ pub mod signed;
 
 #[cfg(feature = "test_utils")]
 pub mod conformance;
-
-pub use access::Access;
-pub use certificate::Certificate;
-pub use delegation::Delegation;
-pub use graph::AuthGraph;
-pub use id::Id;
-pub use keyline::{Inserted, Keyline};
-pub use revocation::Revocation;
-pub use signed::{Signed, Verified};
